Added generic error handling. Used fallback error example from Jflex documentation.
"
The last lexical rule in the example specification is used as an error fallback. It matches any character in any state that has not been matched by another rule.
It doesn’t conflict with any other rule because it has the least priority (because it’s the last rule) and because it matches only one character
(so it can’t have longest match precedence over any other rule).
"